Adding PawnTracker for posterity.
Changing to an easier approach for en passant but just a very bare reminder in
case we want to visit the idea of a PawnTracker in the future.

(My rough estimate is that PawnTracker is lighter in memory without sacrificing
EnPassantFilter speed but at the cost of an extra book keeping step. But at this
point this is all just premature optimization.)